Don't use long for DDS magic number
It's a DWORD (i.e. 32bit int) according to spec, so using long is
actually wrong when sizeof(long) == 8.

+++ neo/renderer/Image_load.cpp
@@ -1400,7 +1400,7 @@ bool idImage::CheckPrecompressedImage( bool fullLoad ) {
 
 	fileSystem->CloseFile( f );
 
-	unsigned long magic = LittleLong( *(unsigned long *)data );
+	unsigned int magic = LittleLong( *(unsigned int *)data );
 	ddsFileHeader_t	*_header = (ddsFileHeader_t *)(data + 4);
 	int ddspf_dwFlags = LittleLong( _header->ddspf.dwFlags );
